+++ src/lib/SimpleHtmlDom/SimpleHtmlDom.php
@@ -1,5 +1,5 @@
 <?php
-namespace topshelfcraft\scraper\lib\SimpleHtmlDom;
+namespace TopShelfCraft\Scraper\lib\SimpleHtmlDom;
 
 require '_functions.php';
 
@@ -9,7 +9,7 @@ class SimpleHtmlDom {
 	 * @return simple_html_dom
 	 */
 	static public function file_get_html() {
-		return call_user_func_array ('\topshelfcraft\scraper\lib\SimpleHtmlDom\file_get_html', func_get_args());
+		return call_user_func_array ('\TopShelfCraft\Scraper\lib\SimpleHtmlDom\file_get_html', func_get_args());
 	}
 
 	/**
@@ -17,7 +17,7 @@ class SimpleHtmlDom {
 	 * @return simple_html_dom
 	 */
 	static public function str_get_html() {
-		return call_user_func_array ('\topshelfcraft\scraper\lib\SimpleHtmlDom\str_get_html' , func_get_args());
+		return call_user_func_array ('\TopShelfCraft\Scraper\lib\SimpleHtmlDom\str_get_html' , func_get_args());
 	}
 
 }