+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Scraper Changelog
 
+The format of this file is based on ["Keep a Changelog"](http://keepachangelog.com/). This project adheres to [Semantic Versioning](http://semver.org/). Version numbers follow the pattern: `MAJOR.FEATURE.BUGFIX`
+
+
+## 4.0.0 - 2022-10-01
+
+### Improved
+
+- Scraper is ready for Craft 4!
+
+### Changed
+
+- Updated Symfony (Goutte) package to 4.x; Some internal methods and return types have changed.
+
+### Removed
+
+- Removed `Scraper::$plugin` static accessor; use `Scraper::getInstance()` instead.
+
+### Deprecated
+
+- Deprecated SimpleHTMLDom driver. In future versions, `scraper.get()` will use a different driver by default. 
+
 ## 3.0.0 - 2020-10-26
 
 ### Changed